feat: emotion 세팅
