+++ apps/web/next.config.mjs
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  compiler: {
+    emotion: true,
+  },
+};
 
 export default nextConfig;
