+++ package.json
@@ -10,11 +10,17 @@
     "lint": "turbo lint",
     "format": "prettier --write \"**/*.{ts,tsx,md}\""
   },
+  "dependencies": {
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5"
+  },
   "devDependencies": {
+    "@emotion/babel-plugin": "^11.11.0",
+    "@emotion/babel-preset-css-prop": "^11.11.0",
+    "next": "14.2.4",
     "prettier": "^3.2.5",
     "turbo": "^2.0.4",
-    "typescript": "^5.4.5",
-    "next": "14.2.4"
+    "typescript": "^5.4.5"
   },
   "packageManager": "pnpm@8.15.6",
   "engines": {
