+++ packages/core/typescript-config/base.json
@@ -15,6 +15,7 @@
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "strict": true,
-    "target": "ES2022"
+    "target": "ES2022",
+    "types": ["@emotion/react/types/css-prop"]
   }
 }
